Refactor AnsiKeyboardParser to be more extensible
